Sun, Jun 26, 2022, 9:38 PM +00:00

+++ pset4/filter-more/helpers.c
@@ -66,9 +66,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                     }
                 }
             }
-            new_image[i][j].rgbtBlue = round((float) tmpBlue / counter);
-            new_image[i][j].rgbtRed = round((float) tmpRed / counter);
-            new_image[i][j].rgbtGreen = round((float) tmpGreen / counter);
+            new_image[i][j].rgbtBlue = round(tmpBlue / counter);
+            new_image[i][j].rgbtRed = round( tmpRed / counter);
+            new_image[i][j].rgbtGreen = round(tmpGreen / counter);
         }
     }
